requirements.txt
blurhash==1.1.4
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==5.1.1
distlib==0.3.8
filelock==3.13.4
idna==3.7
Mastodon.py==1.8.1
platformdirs==4.2.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
requests==2.32.3
six==1.16.0
urllib3==2.2.2
virtualenv==20.25.3
